+++ bootique-jdbc-test/src/main/java/io/bootique/jdbc/test/TestDataManager.java
@@ -1,0 +1,44 @@
+package io.bootique.jdbc.test;
+
+import org.junit.rules.ExternalResource;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * A unit test "rule" to manage data sets for unit tests.
+ *
+ * @since 0.13
+ */
+public class TestDataManager extends ExternalResource {
+
+    private Table[] tablesInInsertOrder;
+    private Map<String, Table> tables;
+
+
+    public TestDataManager(Table... tablesInInsertOrder) {
+        this.tablesInInsertOrder = tablesInInsertOrder != null ? tablesInInsertOrder : new Table[0];
+
+        tables = new HashMap<>();
+        for (Table t : this.tablesInInsertOrder) {
+            tables.put(t.getName(), t);
+        }
+    }
+
+    @Override
+    protected void before() throws Throwable {
+
+        // do cleanup before the test; leave the data alone *after* - perhaps we want to inspect it manually, etc.
+
+        for (int i = tablesInInsertOrder.length - 1; i >= 0; i--) {
+            tablesInInsertOrder[i].deleteAll();
+        }
+    }
+
+    public Table getTable(String name) {
+        return tables.computeIfAbsent(name, n -> {
+            throw new IllegalArgumentException("Unknown table name: " + n);
+        });
+    }
+
+}